WebServer/requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.1.2
django-braces==1.13.0
django-celery-results==1.0.1
django-extensions==2.1.3
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
googlemaps==3.0.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
matcher==0.2
oauthlib==2.1.0
openapi-codec==1.3.2
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.5
redis==2.10.6
requests==2.20.0
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
Unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.24
vine==1.1.4
